corindagpt/requirements.txt
httpx==0.27.*
langchain-core==0.2.*
PyYAML==6.*
SpeechRecognition==3.10.*
pynput==1.7.*
python-dotenv==1.*
elevenlabs>=1.0.0
